+++ tests/Integration/Execution/MutationExecutor/HasManyTest.php
@@ -314,7 +314,7 @@ final class HasManyTest extends DBTestCase
     /**
      * @return array<array<string, string>>
      */
-    public function existingModelMutations(): array
+    public static function existingModelMutations(): array
     {
         return [
             ['Update action' => 'update'],