+++ tests/Integration/Execution/MutationExecutor/HasOneTest.php
@@ -253,7 +253,7 @@ GRAPHQL
     /**
      * @return array<array<string, string>>
      */
-    public function existingModelMutations(): array
+    public static function existingModelMutations(): array
     {
         return [
             ['Update action' => 'update'],