+++ tests/Integration/Execution/MutationExecutor/MorphToTest.php
@@ -180,7 +180,7 @@ final class MorphToTest extends DBTestCase
     /**
      * @return array<array<string, string>>
      */
-    public function existingModelMutations(): array
+    public static function existingModelMutations(): array
     {
         return [
             ['Update action' => 'update'],